Add delete authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -20,7 +22,7 @@ class ListingController
      */
     public function index()
     {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
         loadView('listings/index', ['listings' => $listings]);
     }
     /**
@@ -59,7 +61,7 @@ class ListingController
     {
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
-        $newListingData['user_id'] = 2;
+        $newListingData['user_id'] = Session::get('user')['id'];
         $newListingData = array_map('sanitize', $newListingData);
         $requiredFields = ['title', 'description', 'email', 'city', 'state', 'salary'];
         $errors = [];
@@ -87,10 +89,10 @@ class ListingController
                 $values[] = ':' . $field;
             }
             $values = implode(', ', $values);
-            inspect($values);
             $query = "INSERT INTO listings ($fields) VALUES ($values)";
             $this->db->query($query, $newListingData);
             // Redirect to listings
+            $_SESSION['success_message'] = 'Listing added successfully';
             redirect('/listings');
         }
     }
@@ -108,9 +110,15 @@ class ListingController
             'id' => $id
         ];
         $listing = $this->db->query('SELECT * FROM listings WHERE id=:id', $params)->fetch();
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing!';
+            return redirect('/listings/' . $listing->id);
         }
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
         // Set flash message
